The GetNextMatchIntent has been updated to better handle the getting information from Alexa. When parsing the response it now looks through the ambiguity resolutions parameter and checks for a successful match.
Also, the intent can now detect if a match is happening currently for the
requested team and reports the status of the current game along with the
requested next match information.

Added a few arrays of phrases which are randomly selected to fill the output
speech response. These should be ported in the future to language strings
to be more robust in the app.

Some TODOs are listed throughout.