Adding a sort of intermediate commit to log code changes for if/when we may want to revert and have an example of what to do.
This commit shows how to implement an audio player device as well as
how to embedd output audio as part of the ssml speech output. Since
our audio clips are short it is more convenient and provides a better
experience if the clips are quickly answered.

But we may want to move to using an audio player later for some reason
and this version also started to implement the required Amazon intents
for using the audioPlayer object and may provide some help in the
future.