The index is cleaned up a little more with some of the resources abstracted to an untracked resources.js to protect keys and urls etc.
Also this commit fixes the launch request sequence waiting to
correctly listen for input and an updated welcome message to
be more descriptive about current commands. For fun, this
welcome message also includes audio embedded in the ssml using
the audio tag. This adds another level to the app without having
to use an audio player since the audio is short. See previous
commits for an example of the full audio player if we need to
updated and add full audio in the future.