Cleaned up code from last commit to make the structure more uniform and hopefully more understandable.
The major feautre of the last commit was introducing the callback
structure and flow to handle multiple api calls. The idea is to
use the attributes member of the event object to store information
as different apis are traversed. Along the way some apis may need
to stop to 'catch a breath' and save information if necessary.

At the beginning of each intent the callback strucutre is defined
and is implemented as a stack so it is a first on first off
approach and so the very first entry in the array is where the
skill should land at the very on with all of the required info
and produce a response to alexa. As mentioned along the way
the api may need to land to handle parsing a response and then
continue on. This can either be handled by following the same
api call method or just popping the next callback and continuing.

This commit also incoroporates the getNextStage intent. Simply
responding what stage OWL is currently in. This could easily be
extended to also report how many matches have been played or what
week we are in, etc.