Fixed problems from our most recent submission. Forgot to handle the case for when users ask for teams greater than 12.
Also fixed the getNextMatchIntent regardless of the team. This would
have failed tonight if not caught. We were comparing the wrong start
time. When working with standings therea are now two time fileds. The
filed startDate which contains the time string (this is different than
the getNextTeamMatch wherte startDate contains the time since 1970 in
seconds) and then there is startDateTS which contains the time since
1970 seconds that we need.

Also fixed the live match time and status reporting. There was a bug
where the isTied status would be overwritten by isLosing status message.

Also correctly checked the next match for any team and a specific team
to say today or tomrrow. The correct thing to do was to add the timezone
offset to the nowSeconds generated in the get calendar time function.