Extend deserializer
Added ability to detect models which have a class attribute that is a list of
other model objects or builtin types. To complete this also needed to add
functionality to import classes that lived outside the current module and we
only knew about by their name as a str (e.g., 'owl_model.player.Player').