Extend Model Definitions
Looked at a response from the /teams endpoint and extended the model to include
many of the attributes we would possibly be interested in now and in the future.

Began to add a owl_model.URL class. The idea here was that the URL class when
instantiated could go and fetch the contatct or have a wrapper method that could
handle the fetch of data without the user having to explicitly making the
request. AS of now the URL class is a place holder as there is no model defined
or way to handle it in the deserializer because mapping a URL to all the
different names that could be defined in the response from OWL (e.g., icon,
logo, accounts, twitter, etc.) needs to be figured out.