Extend serialier and add url class
The approach with the URL class is to provide flexibility and have class
behavior but still abstract the behavior of the url to a parent class. However,
the way the OWL API returns responses does not represent these URLs is a way to
discriminate at creation time the class. For example, the URL is not in the
response payload as
  { ..., 'url': {'logo': 'https://...'}, ... }
instead it is
  { ..., 'logo': 'https://...', ... }.
This means if we want the URL to be a class of type logo it doesn' match the
serializers method to parse the object descending an object tree in an object
notation way. As a result the response payload is a primitive data type when we
are expecting to parse an object model.

The way the serializer overcomes this is in the class implementation the
bootstrap_subclass method will promote the primitive data type, a str in this
case, to a dictionary and then continue.

There is a todo mentioned that suggests doing the same thing but a slightly
different approach where the parent class contains a third dictionary similar to
the attribute mapping dictionary to map the payload to the object that should be
created acting as sort of a discriminator map.

Some other clean up to the code. Modified the base modelobject class to include
printing the name of the class because it sometimes would trip me up to see a
dictionary prinited even though that was the expected behavior.