Extend deserializer and add Match class
The deserializer can now parse date strings in ISO format and converts them to
datetimes. This also includes timestamps expressed as integers in milliseconds
since epoch. The deserializer should now be minimally complete to handle
everything required by the model classes.

I made a note at the end of the deserialize method that propses including
another function to the model objects which allow for further initialization
after the object has been created and configured with all it's information. This
in my opinion is an interesting and useful idea as it can further isolate steps
that someone would need to do to make the class more useful. A concrete example
would be with applying timezones to the match dates so that the correct time can
be determined.

Added the match class. We should be to the point of just further adding model
classes that build up to match the expected response from the OWL API.