Modify deserializer
Working with the models and trying to build class up to the basic API request
noticed the desrializer could be improved by allowing lists of dictonaries. This
wasn't possible before because the .split(',') method would split on all ','
characters and if a dict( , ) were present it would split that in half breaking
the deserializer. A regex has been added to modify that.

This change however may not actually be used because in the following model
commit while this change helps the deserializer be robust, in this specific case
it made the model a little awkward by having to access the teams in the
TeamRequest like a dictionary. But, the functionality should be there anyway for
when/if needed later.