Extend model and add new model objects
This commit finishes a minimal working model for everything needed by a request
to the OWL API at the /teams endpoint.

The apirequst class is still very much a fluid class. It is still a little fuzzy
to me how to best use the class. For example, should it extend the url.URL class
or should it stand by itself? I am leaning to having the url.URL class be
involved, but we will see.

The Map object has some static class functionality that are more Overwatch game
mechanic related but I think actually helps mature the class and the entire
model as a whole. The Team object will most likely include all teams in the
league as well.

The finalize_init method has also made it into the model as I started to find
more and more use cases that make it effective to have. This was also added to
the deserializer in the previous commit. So there are now two different methods
that affect the control flow and construction of an object. There is the
'bootstrap_subclass' method and the 'finalize_init' method.

'bootstrap_subclass' can be used to massage and modify the data payload in the
deserializer to help the subclass initialize and setup payload parameters as
expected. Then the 'finalize_init' method is the last method that is called, if
present, to pull together any required information to prepare the class for use.