Work on APIRequest
With OWL starting soon trying to get a working version running. But since the
model is now getting closer to the interface level there are some questions on
the approach and for now a design direction needs to be made and we can evaluate
down the road how to adjust, if needed.This commit is therefore a snapshot of
the current state with issues/todo's noted to look back at.

The approach now will be to seperate the APIRequest from the returned classes
that will be used. For example, in this snapshot look at how the TeamsRequest
and TeamsByIdRequest is done. This directino will be headed more towards the
TeamsByIdRequest where the makeapicall() method will return a class in the model
instead of how the TeamsByIdRequest returns a TeamsRequest method.

This has the benefit to remove some of the api methods from the model classes
and keeping things seperated. There fore a League model class will be created
for the TeamsRequest to operate on since the closest thing this request returns
in the OWL model paradigm is a League as it returns a collection of all the
teams.

I am also noticing some of the bootstraping may be getting complicated and
over the top. Because the idea was to write models that conformed to the
structure of the API. However as noted in the Team bootstrap method (where this
started to be a concern) the OWL API is inconsistent with the information
provided. So to some degree a bootstrap method is required and in a general
OWL independent deserializer it is kind of a nice feature to have.