APIRequest ready for use
After messing around with the APIRequest class it is in a state that is usaable
and I feel pretty trasparent.

To overcome many of the design questions the end result is that the APIRequest
is the umbrella factory class with classmethods that are the factory methods
making the api calls and returning objects from the owl_model library.

The main block of apirequest.py has the beginning of some examples for how to
use the APIRequest class and how to begin to utilize the information from the
deserializer.