Rearrange lambda skill into functional units
The goal here is to break the lambda function up into smaller logical parts that
provides more intuition for how the ASK SDK is set up.

The skill builder is configured in lambda.py and all of the request handlers,
exception handlers and interceptors inside owl_skill are added here.