Working GetNextMatchIntent
Local tests indicate GetNextTeamMatchIntent is ready to test on Alexa Simulator
and hardware.

There is still more that can be done with this intent. For example, the V1 SDK
implementation checked if a live match was happening and as of right now this
version does not. While still helpful to check a live match it is worth
considering the approach (i.e., just translate the old code or is there a
smarter way?).