Implement the GetTeamRecordIntent
The old endpoint we were using last year seems to have a bug where every
teams records section is the same being updated with the team that is in
first place. Digging into the OWL API a little more there is a /v2/teams
endpoint (along with other /v2/ calls) that has the correct information.
However, the format is a little different than what we have building up
to. The owl_model.team.Team was slighlty modified to support a call to
/v2/teams for a single team id but in general wouldn't work if called on
all teams. Therefore it was not added to owl_model.apirequest to fetch
all teams.